refactored location of the markdown component
